Import DHS data from API

Read a list of raster data, aggregate by shapefile. This function is designed for the raster data with a third dimension called year. For example, we have a list of raster data, each raster data is a geotiff file, and the file name is like this: pfpr_2to10_2000.tif, pfpr_2to10_2001.tif, pfpr_2to10_2002.tif, etc. The smart_read_files_by_year function will read all the files in the list and return a list of raster data, each raster data is a star object, and the star object has a third dimension called year, which is the year of the raster data. Internally, this function use the stars package to read the raster data. The stars package is a very powerful package for reading and manipulating raster data. It is much faster than the raster package. The package use stars proxy object to speed up the reading process. More information about the stars package can be found here: https://r-spatial.github.io/stars/

Provide a stars or stars-proxy object, a shapefile and a function for aggregation, this function will help you extract the raster data within the shapefile, and then merge it back. A very useful function to extract GIS data.

Provide a stars or stars-proxy object, a shapefile, this function will help you crop the raster data within the shapefile very useful function to extract GIS data.

This function is a smart file reader, it will detect the file path and return the file list with dates. For example, we have a file path like this: 'C:/Users/username/Desktop/Bo District/Bo District 2020-09.xlsx' 'C:/Users/username/Desktop/Bo District/Bo District 2020-10.xlsx' 'C:/Users/username/Desktop/Bo District/Bo District 2020-11.xlsx' ... If we want to read all the files in the folder and extract the date from the file name, we can use this function to read all the files. The function will detect the patterns in the file path and return the file list with dates. The patterns should be like this: Year pattern: yyyy Month pattern: mm Day pattern: dd

Dissect data by admin level and date level TODO: next step, should be able to include extra column in the data.table, such as hf_type

Usage

import_dhs(year = 2021)

raster_read_aggregate_year(smart_list, shp, fun, index)

star_aggregate(star, shp, fun)

stars_crop(star, shp)

bind_who_gis_adm2_shp(
  shp_regional_afro_adm2,
  shp_regional_amro_adm2,
  shp_regional_emro_adm2,
  shp_regional_euro_adm2,
  shp_regional_searo_adm2,
  shp_regional_wpro_adm2,
  shp_global_adm2
)

get_files(smart_path)

dissect_meta(d)

Arguments

year

DHS year to be imported

smart_list

a list of raster data should be a list read from the smart_read functions for example, we have a list of raster data, each raster data is a geotiff file, and the file name is like this: pfpr_2to10_2000.tif, pfpr_2to10_2001.tif, pfpr_2to10_2002.tif, etc. The smart_read_files_by_year function will read all the files in the list and return a list of raster data, each raster data is a star object, and the star object has a third dimension called year, which is the year of the raster data. The smart list should be returned by the smart_read functions. It is a list with two elements: files and years. The files element is a list of file paths, and the years element is a list of years. More information can be found in the smart_read functions.

shp

a shapefile object

fun

a function for aggregation

index

a name for the index, for example, pfpr, rainfall, etc. after aggregation, the function will return a shapefile with the aggregated data, and the column name will be the index name. If we don't provide the index name, the column name will be the file name of the raster data. Which is not very useful.

star

a star object

shp_regional_afro_adm2

Path to the shapefile of the WHO African Region

shp_regional_amro_adm2

Path to the shapefile of the WHO American Region

shp_regional_emro_adm2

Path to the shapefile of the WHO Eastern Mediterranean Region

shp_regional_euro_adm2

Path to the shapefile of the WHO European Region

shp_regional_searo_adm2

Path to the shapefile of the WHO South-East Asia Region

shp_regional_wpro_adm2

Path to the shapefile of the WHO Western Pacific Region

shp_global_adm2

Path to the shapefile of the WHO Global Region

smart_path

A string, the path of the file. Should contain patterns defined above.

d

Data table

Value

A tibble of DHS data

a shapefile with the aggregated data

a shapefile with the aggregated data

a cropped stars object

A shapefile of the WHO Global Region

A list of file path and along with a date list. The result example: $files 1 'C:/Users/username/Desktop/Bo District/Bo District 2020-09.xlsx' 2 'C:/Users/username/Desktop/Bo District/Bo District 2020-10.xlsx' 3 'C:/Users/username/Desktop/Bo District/Bo District 2020-11.xlsx' $dates 1 '2020-09-01' '2020-10-01' '2020-11-01'

Details

Day pattern should be used with year and month pattern. If the day pattern is found, then both year and month pattern should be found. Otherwise, the function will throw an error. The same rule applies to month pattern. If the month pattern is found, then year pattern should be found. Otherwise, the function will throw an error.

Author

Chunzhe ZHANG